solar_model/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.5.0
scipy>=1.7.0
torch>=1.9.0  # Required for AgentTorch
torchvision
torchaudio
agent-torch>=0.1.0  # Main AgentTorch package